pyproject.toml
[tool.poetry]
name = "crawlkit"
version = "0.1.0"
description = "Crawlkit — offline-friendly research, extraction, and compliance toolkit."
authors = ["ACES Aerodynamics Analyst <info@acesaero.co.za>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "crawlkit" }, { include = "watercrawl" }]
exclude = [
  "apps",
  "codex",
  "data",
  "data_contracts",
  "docs",
  "docs-starlight",
  "examples.py",
  "main.py",
  "node_modules",
  "platform",
  "presets",
  "scripts",
  "tests",
  "tools",
]

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
pandas = "2.3.3"
openpyxl = "^3.1.5"
requests = "2.32.5"
httpx = "0.28.1"
tenacity = "^9.1.2"
pydantic = "^2.12.3"
structlog = "25.4.0"
prometheus-client = "0.23.1"
pydantic-settings = "^2.11.0"
boto3 = "^1.40.57"
azure-identity = "^1.25.1"
azure-keyvault-secrets = "^4.10.0"
pint = "0.25"
scrapy = "2.13.3"
scrapy-playwright = "^0.0.44"
playwright = "^1.55.0"
selectolax = "^0.4.0"
trafilatura = "^2.0.0"
extruct = "^0.18.0"
markdownify = "^1.2.0"
phonenumbers = "^9.0.16"
dnspython = "^2.8.0"
pyyaml = "^6.0.3"
opentelemetry-api = "^1.38.0"
opentelemetry-sdk = "^1.38.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.38.0"
opentelemetry-exporter-prometheus = "^0.59b0"
marshmallow = { version = ">=3.26.1,<4", optional = true }

[tool.poetry.group.contracts]
optional = true

[tool.poetry.group.contracts.dependencies]
great-expectations = { version = "1.7.1", python = ">=3.9,<3.14" }

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
pytest-asyncio = "1.2.0"
pytest-cov = "7.0.0"
ruff = "0.14.1"
black = "25.9.0"
isort = "7.0.0"
mypy = "1.18.2"
bandit = "1.8.6"
dotenv-linter = "0.7.0"
pre-commit = "4.3.0"
pandas-stubs = "*"
types-requests = "2.28.11"
jupyter = "1.1.1"
hypothesis = "6.142.3"
pydocstyle = "^6.3.0"
safety = "^3.6.2"
sqlfluff = "^3.5.0"
sqlfluff-templater-dbt = "^3.5.0"
yamllint = "^1.19.0"
yamlfix = "^1.16.0"
safety-db = "^2021.7.17"
types-boto3 = "1.40.57"
types-botocore = "1.0.2"
types-pillow = "*"
types-pyyaml = "6.0.12"
types-setuptools = "*"
types-urllib3 = "1.26.25"
types-python-dateutil = "2.8.19"
types-networkx = "3.5.0.20251001"
narwhals = "2.9.0"
streamlit = "^1.50.0"
axe-selenium-python = "^2.1.6"
selenium = "^4.37.0"
mutmut = "3.3.1"
tomli = "^2.3.0"

[tool.poetry.group.dbt]
optional = true

[tool.poetry.group.dbt.dependencies]
# dbt dependencies are heavy and platform-specific. Keep them in an optional group
# so the baseline dev environment can resolve on machines where dbt-core 1.11.0
# (or platform wheels) are not available. To enable locally: `poetry install --with dbt`.
"dbt-core" = { version = "1.10.13", python = ">=3.13,<3.15" }
dbt-duckdb = { version = "^1.9.6", python = ">=3.13,<3.15" }

[tool.poetry.group.ui]
optional = true

[tool.poetry.group.ui.dependencies]
streamlit = "^1.50.0"
pyarrow = { version = "21.0.0", python = ">=3.13,<3.14" }

[tool.poetry.group.lakehouse]
optional = true

[tool.poetry.group.lakehouse.dependencies]
deltalake = { version = "1.2.1", python = ">=3.13,<3.14" }

[tool.poetry.extras]
pipeline = ["marshmallow"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312", "py313", "py314"]
include = "\\.pyi?$"
extend-exclude = '''
(
  ^/node_modules/
  |^/docs-starlight/
)
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["apps", "crawlkit", "watercrawl", "scripts"]
skip_glob = [".trunk/plugins/**/*"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["apps", "crawlkit", "watercrawl", "scripts", "tests"]
extend-exclude = ["stubs/third_party"]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E203", "E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "I001"]
"watercrawl/integrations/contracts/great_expectations_runner.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
filterwarnings = [
  # Suppress Great Expectations Pydantic V1 warning since GE is pinned to 1.7.1 for Python < 3.14
  "ignore:Core Pydantic V1 functionality isn't compatible with Python 3.14 or greater:UserWarning",
  # Allow other warnings to show
]

[tool.mypy]
python_version = "3.13"
show_error_codes = true
mypy_path = ["stubs/third_party", "stubs"]
exclude = ["^stubs/third_party/"]

[[tool.mypy.overrides]]
module = [
  "boto3",
  "botocore.*",
  "azure.identity",
  "azure.keyvault.secrets",
  "duckdb",
  "great_expectations",
  "great_expectations.*",
  "pyspark",
  "pyspark.*",
  "whylogs",
  "whylogs.*",
  "dns",
  "dns.*",
  "celery",
  "fastapi",
]
ignore_missing_imports = true

[tool.mutmut]
paths_to_mutate = [
  "watercrawl/application/pipeline.py",
  "watercrawl/integrations/adapters/research/core.py",
]
tests_dir = "tests"
runner = "python -m pytest tests/test_pipeline.py tests/test_research_logic.py"
